reservation controller

+++ src/main/java/com/laioffer/staybooking/service/StayService.java
@@ -1,11 +1,10 @@
 package com.laioffer.staybooking.service;
 
+import com.laioffer.staybooking.exception.StayDeleteException;
 import com.laioffer.staybooking.exception.StayNotExistException;
-import com.laioffer.staybooking.model.Location;
-import com.laioffer.staybooking.model.Stay;
-import com.laioffer.staybooking.model.StayImage;
-import com.laioffer.staybooking.model.User;
+import com.laioffer.staybooking.model.*;
 import com.laioffer.staybooking.repository.LocationRepository;
+import com.laioffer.staybooking.repository.ReservationRepository;
 import com.laioffer.staybooking.repository.StayRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,12 +25,15 @@ public class StayService {
     private LocationRepository locationRepository;
     private GeoCodingService geoCodingService;
 
+    private ReservationRepository reservationRepository;
+
     @Autowired
-    public StayService(StayRepository stayRepository, ImageStorageService imageStorageService, LocationRepository locationRepository, GeoCodingService geoCodingService) {
+    public StayService(StayRepository stayRepository, ImageStorageService imageStorageService, LocationRepository locationRepository, GeoCodingService geoCodingService, ReservationRepository reservationRepository) {
         this.stayRepository = stayRepository;
         this.imageStorageService = imageStorageService;
         this.locationRepository = locationRepository;
         this.geoCodingService = geoCodingService;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<Stay> listByUser(String username) {
@@ -68,6 +70,10 @@ public class StayService {
         Stay stay = stayRepository.findByIdAndHost(stayId, new User.Builder().setUsername(username).build());
         if (stay == null) {
             throw new StayNotExistException("Stay doesn't exist");
+        }
+        List<Reservation> reservations = reservationRepository.findByStayAndCheckoutDateAfter(stay, LocalDate.now());
+        if (reservations != null && reservations.size() > 0) {
+            throw new StayDeleteException("Cannot delete stay with active reservation");
         }
         stayRepository.deleteById(stayId);
     }
